Moves Client to actor model
This moves the `Client` to an `actor` model as the `Token` is owned state and authorization can happen on any thread. Also gets rid of the publisher in favor of a more straightforward authorization model. This means clients might have to do a bit more groundwork (or not, depending on usage). Updated the sample app and README as documentation for the new methods.